requirements.txt
axios
react-router-dom
flask
flask-cors
mongodb
Pymongo
Flask-Pymongo
python-dotenv
openai
pandas
pymupdf